refactor: auth 도메인 구현, token dto 생성
