+++ marketplace/src/main/java/store/carjava/marketplace/app/auth/dto/TokenRequest.java
@@ -1,0 +1,7 @@
+package store.carjava.marketplace.app.auth.dto;
+
+public record TokenRequest(
+        String authorizationCode
+) {
+
+}
