+++ marketplace/src/main/java/store/carjava/marketplace/app/auth/dto/TokenResponse.java
@@ -1,0 +1,11 @@
+package store.carjava.marketplace.app.auth.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public record TokenResponse(
+        @JsonProperty("access_token") String accessToken,
+        @JsonProperty("refresh_token") String refreshToken
+) {
+
+}
+
