--- marketplace/src/main/java/store/carjava/marketplace/web/TokenRequest.java
@@ -1,14 +1,0 @@
-package store.carjava.marketplace.web;
-
-public class TokenRequest {
-    private String authorizationCode;
-
-    // Getter and Setter
-    public String getAuthorizationCode() {
-        return authorizationCode;
-    }
-
-    public void setAuthorizationCode(String authorizationCode) {
-        this.authorizationCode = authorizationCode;
-    }
-}
