--- marketplace/src/main/java/store/carjava/marketplace/web/TokenResponse.java
@@ -1,28 +1,0 @@
-package store.carjava.marketplace.web;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public class TokenResponse {
-    @JsonProperty("access_token")
-    private String accessToken;
-
-    @JsonProperty("refresh_token")
-    private String refreshToken;
-
-    // Getters and Setters
-    public String getAccessToken() {
-        return accessToken;
-    }
-
-    public void setAccessToken(String accessToken) {
-        this.accessToken = accessToken;
-    }
-
-    public String getRefreshToken() {
-        return refreshToken;
-    }
-
-    public void setRefreshToken(String refreshToken) {
-        this.refreshToken = refreshToken;
-    }
-}
