--- marketplace/src/main/resources/templates/login/callback.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Authorization Code</title>
-</head>
-<body>
-<h1>Authorization Code</h1>
-<p>Your authorization code is:</p>
-<pre th:text="${authorizationCode}">Authorization Code will appear here</pre>
-</body>
-</html>
