--- marketplace/src/main/resources/templates/login/index.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-    <title>Login</title>
-</head>
-<body>
-<h1>Welcome to the Application</h1>
-<p>Click the button below to log in:</p>
-<form action="/login/redirect" method="get">
-    <button type="submit">Login with Keycloak</button>
-</form>
-</body>
-</html>
